ex5/requirements.txt
# Context Windows Lab - Requirements
# Python 3.8+ required

# Core LLM and ML Libraries
langchain>=0.1.0
chromadb>=0.4.0
ollama>=0.1.0
sentence-transformers>=2.2.0

# Data Science and Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text Processing
Faker>=20.0.0
python-Levenshtein>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Optional: Enhanced visualization
plotly>=5.17.0
